Rewrite orbiter without need for THREE
The orbiter headlining my current site requires three, which is a huge
import. This rewrites the orbiter, with some reorganization as well,
managing the graphics manually, with the help of Sylvester.

The added advantage is that the generality is already there for "softening"
the points under rotation, so long as they are puppeteered by a custom
Geometry.